requirements.txt
# For OAuth and HTTP requests
requests
python-dotenv

# For audio feature extraction
librosa
numpy
scipy
soundfile
pytube
yt_dlp

# For audio previews
pydub

# For plotting/visualizations (optional but useful)
matplotlib
seaborn

# Backend framework
fastapi
uvicorn

# Optional: for building a Streamlit UI
streamlit
